+++ tests/run-almost-all.ts
@@ -230,6 +230,7 @@ import { testit as issue6127 } from "./issue-6127.ts";
 import { testit as issue6163 } from "./issue-6163.ts";
 import { testit as issue6166 } from "./issue-6166.ts";
 import { testit as issue6169 } from "./issue-6169.ts";
+import { testit as issue5771 } from "./issue-5771.ts";
 import { testit as issue5980 } from "./issue-5980.ts";
 import { testit as issue5982 } from "./issue-5982.ts";
 import { testit as issue6035 } from "./issue-6035.ts";
@@ -370,6 +371,7 @@ export const tests: NamedTest[] = [
   ["create-annot-enters-edit-mode", createAnnotEntersEditMode],
   ["issue-6166", issue6166],
   ["issue-6169", issue6169],
+  ["issue-5771", issue5771],
   ["issue-6111", issue6111],
   ["annot-list-placement", annotListPlacement],
   ["exit-edit-pdf-deselects", exitEditPdfDeselects],